feat: delete clients and projects

+++ src/main/java/dev/ravel/proflow/domain/services/ClientServiceImpl.java
@@ -1,6 +1,8 @@
 package dev.ravel.proflow.domain.services;
 
 import dev.ravel.proflow.infrastructure.model.Client;
+import dev.ravel.proflow.infrastructure.model.Project;
+import dev.ravel.proflow.infrastructure.model.Task;
 import dev.ravel.proflow.infrastructure.repository.ClientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,9 +12,11 @@ import java.util.List;
 @Service
 public class ClientServiceImpl implements ClientService {
     private final ClientRepository clientRepository;
+    private final ProjectService projectService;
 
-    public ClientServiceImpl(ClientRepository clientRepository) {
+    public ClientServiceImpl(ClientRepository clientRepository, ProjectService projectService) {
         this.clientRepository = clientRepository;
+        this.projectService = projectService;;
     }
 
     @Override
@@ -32,7 +36,16 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public void deleteClient(long id) {
-        clientRepository.deleteById(id);
+        Client client = clientRepository.findById(id).orElse(null);
+        if (client != null) {
+            List<Project> projects = projectService.getProjectsByClientId((int) id);
+            if (projects != null) {
+                projects.forEach(project -> {
+                    projectService.deleteProject((long)project.getId());
+                });
+            }
+            clientRepository.deleteById(id);
+        }
     }
 
     @Override
